Release 2.3.0

+++ modules/local/itsx_cutasv.nf
@@ -8,10 +8,14 @@ process ITSX_CUTASV {
 
     input:
     path fasta
+    val outfile
 
     output:
-    path "ASV_ITS_seqs.full.fasta", emit: fasta
-    path "versions.yml"          , emit: versions
+    path outfile         , emit: fasta
+    path "ASV_ITS_seqs.summary.txt", emit: summary
+    path "ASV_ITS_seqs.*fasta", emit: fastas
+    path "versions.yml"  , emit: versions
+    path "*.args.txt"    , emit: args
 
     script:
     def args = task.ext.args ?: ''
@@ -22,6 +26,12 @@ process ITSX_CUTASV {
         --cpu $task.cpus \\
         -o ASV_ITS_seqs
 
+    if [ ! -s $outfile ]; then
+        echo "ERROR: No ITS regions found by ITSx. You might want to modify --cut_its and/or --its_partial" >&2
+        exit 1
+    fi
+
+    echo -e "ITSx\t$args" > ITSx.args.txt
     cat <<-END_VERSIONS > versions.yml
     "${task.process}":
         ITSx: \$( ITSx -h 2>&1 > /dev/null | tail -n 2 | head -n 1 | cut -f 2 -d ' ' )
